# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet API",
        "description": "Documentación de la API de Wallet.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cbu": {
            "get": {
                "summary": "Listar los CBUs y titulares guardados por el usuario autenticado",
                "operationId": "d599a7b440d3b2922875adba0375abed",
                "responses": {
                    "200": {
                        "description": "Listado de cuentas guardadas obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cbu": {
                                                "type": "string",
                                                "example": "1255774332638486003143"
                                            },
                                            "titular": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cbu/{cbu}/users/{idUser}": {
            "post": {
                "summary": "Guardar una cuenta (CBU) en la agenda del usuario",
                "operationId": "c1dcf604edd2000a3b4e1da7c995ab8f",
                "parameters": [
                    {
                        "name": "cbu",
                        "in": "path",
                        "description": "CBU de la cuenta a agendar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1255774332638486003143"
                        }
                    },
                    {
                        "name": "idUser",
                        "in": "path",
                        "description": "ID del usuario autenticado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuenta guardada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuenta guardada exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "El ID de la ruta no coincide con el del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No puedes modificar la lista de otro usuario."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El CBU ingresado no existe en el sistema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La cuenta no existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de negocio (intentar guardar la cuenta propia o una ya guardada)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La cuenta ya está guardada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Eliminar un CBU de la agenda del usuario",
                "operationId": "6da932ee56d36c8d49ba27ab510b8085",
                "parameters": [
                    {
                        "name": "cbu",
                        "in": "path",
                        "description": "CBU de la cuenta a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1255774332638486003143"
                        }
                    },
                    {
                        "name": "idUser",
                        "in": "path",
                        "description": "ID del usuario autenticado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CBU removido de la agenda exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CBU removido exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "El ID de la ruta no coincide con el del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La cuenta no se encuentra en tu lista de guardados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No autorizado."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}